+++ src/core/persistence/migrations/version_2.rs
@@ -2,8 +2,6 @@ use crate::prelude::*;
 use rusqlite::Connection;
 
 pub fn migrate(db: &Connection) -> Result<()> {
-    info!("Deleting all readings because of the changed serialization…");
-
     // language=sql
     db.execute_batch(
         r#"